2.1 Store dir improvement

+++ src/main/java/org/bonitasoft/store/artifactdeploy/DeployStrategyProcess.java
@@ -1,4 +1,4 @@
-package org.bonitasoft.store.artefactdeploy;
+package org.bonitasoft.store.artifactdeploy;
 
 import org.bonitasoft.engine.bpm.process.ActivationState;
 import org.bonitasoft.engine.bpm.process.ProcessActivationException;
@@ -18,8 +18,11 @@ import org.bonitasoft.store.toolbox.LoggerStore;
 
 public class DeployStrategyProcess extends DeployStrategy {
 
-    protected static BEvent EventCantRemoveCurrentProcess = new BEvent(DeployStrategyProcess.class.getName(), 1, Level.APPLICATIONERROR, "Current process can't be removed", "To deploy the new process (same name, same version), current process has to be removed. The operation failed.",
+    protected final static BEvent EventCantRemoveCurrentProcess = new BEvent(DeployStrategyProcess.class.getName(), 1, Level.APPLICATIONERROR, "Current process can't be removed", "To deploy the new process (same name, same version), current process has to be removed. The operation failed.",
             "Deployment of the new process is not possible", "Check the exception");
+
+    protected final static BEvent EventProcessNotLoaded = new BEvent(DeployStrategyProcess.class.getName(), 2, Level.APPLICATIONERROR, "Process is not loaded", "Proces has to be loaded first.",
+            "Deployment of the new process is not possible", "Check the error");
 
     /* *********************************************************************** */
     /*                                                                         */
@@ -35,13 +38,14 @@ public class DeployStrategyProcess extends DeployStrategy {
             processDefinitionId = bonitaAccessor.processAPI.getProcessDefinitionId(process.getName(), process.getVersion());
 
         } catch (ProcessDefinitionNotFoundException pe) {
+            // nothing to do here
         }
 
         if (processDefinitionId != null) {
             try {
                 ProcessDeploymentInfo processDeploymentInfo = bonitaAccessor.processAPI.getProcessDeploymentInfo(processDefinitionId);
-                deployOperation.presentDateArtefact = processDeploymentInfo.getDeploymentDate();
-                deployOperation.presentVersionArtefact = process.getVersion();
+                deployOperation.presentDateArtifact = processDeploymentInfo.getDeploymentDate();
+                deployOperation.presentVersionArtifact = process.getVersion();
                 if (processDeploymentInfo.getDeploymentDate().equals(process.getDate())) {
 
                     process.bonitaBaseElement = bonitaAccessor.processAPI.getProcessDefinition(processDeploymentInfo.getProcessId());
@@ -64,8 +68,8 @@ public class DeployStrategyProcess extends DeployStrategy {
         try {
             processDefinitionId = bonitaAccessor.processAPI.getLatestProcessDefinitionId(process.getName());
             ProcessDeploymentInfo processDeploymentInfo = bonitaAccessor.processAPI.getProcessDeploymentInfo(processDefinitionId);
-            deployOperation.presentDateArtefact = processDeploymentInfo.getDeploymentDate();
-            deployOperation.presentVersionArtefact = processDeploymentInfo.getVersion();
+            deployOperation.presentDateArtifact = processDeploymentInfo.getDeploymentDate();
+            deployOperation.presentVersionArtifact = processDeploymentInfo.getVersion();
 
             // event if the same process was deployement AFTER, we considere this artefact as a NEWVERSION
             deployOperation.detectionStatus = DetectionStatus.NEWVERSION;
@@ -86,35 +90,40 @@ public class DeployStrategyProcess extends DeployStrategy {
         DeployOperation deployOperation = new DeployOperation();
         deployOperation.deploymentStatus = DeploymentStatus.NOTHINGDONE;
 
-        // artefact is the process
-        boolean doLoad = false;
-
-        try {
-            Long processDefinitionId = bonitaAccessor.processAPI.getProcessDefinitionId(process.getName(), process.getVersion());
-            ProcessDeploymentInfo processDeploymentInfo = bonitaAccessor.processAPI.getProcessDeploymentInfo(processDefinitionId);
-
-            if (processDeploymentInfo.getActivationState() == ActivationState.ENABLED)
-                bonitaAccessor.processAPI.disableProcess(processDefinitionId);
-
-            bonitaAccessor.processAPI.deleteProcessDefinition(processDefinitionId);
-            doLoad = true;
-
-        } catch (ProcessDefinitionNotFoundException e) {
-            doLoad = true; // this not exist
-        } catch (DeletionException e) {
-            deployOperation.listEvents.add(new BEvent(EventCantRemoveCurrentProcess, e, "Process Name[" + process.getName() + "] Version[" + process.getVersion() + "]"));
-            doLoad = false;
-        } catch (ProcessActivationException e) {
-            deployOperation.listEvents.add(new BEvent(EventCantRemoveCurrentProcess, e, "Process Name[" + process.getName() + "] Version[" + process.getVersion() + "]"));
+        // artifact is the process
+        boolean doLoad = true;
+        // if the artifact is not loaded, stop now
+        if (!process.isLoaded()) {
+            deployOperation.listEvents.add(new BEvent(EventProcessNotLoaded, "Process Name[" + process.getName() + "] Version[" + process.getVersion() + "]"));
             doLoad = false;
         }
 
+        // first, delete the current one
+        if (doLoad) {
+            try {
+                Long processDefinitionId = bonitaAccessor.processAPI.getProcessDefinitionId(process.getName(), process.getVersion());
+                ProcessDeploymentInfo processDeploymentInfo = bonitaAccessor.processAPI.getProcessDeploymentInfo(processDefinitionId);
+
+                if (processDeploymentInfo.getActivationState() == ActivationState.ENABLED)
+                    bonitaAccessor.processAPI.disableProcess(processDefinitionId);
+
+                bonitaAccessor.processAPI.deleteProcessDefinition(processDefinitionId);
+                doLoad = true;
+
+            } catch (ProcessDefinitionNotFoundException e) {
+                doLoad = true; // this not exist
+            } catch (DeletionException | ProcessActivationException e) {
+                deployOperation.listEvents.add(new BEvent(EventCantRemoveCurrentProcess, e, "Process Name[" + process.getName() + "] Version[" + process.getVersion() + "]"));
+                doLoad = false;
+            }
+        }
         if (doLoad) {
 
-            // deploy it
             try {
                 // bonitaAccessor.processAPI.deployAndEnableProcess(businessArchive);
                 ProcessDefinition processDefinition = bonitaAccessor.processAPI.deploy(((ArtifactProcess) process).getBusinessArchive());
+                deployOperation.deploymentStatus = DeploymentStatus.LOADED;
+
                 bonitaAccessor.processAPI.enableProcess(processDefinition.getId());
 
                 process.bonitaBaseElement = processDefinition;
@@ -122,18 +131,23 @@ public class DeployStrategyProcess extends DeployStrategy {
 
             } catch (ProcessDeployException e) {
                 deployOperation.deploymentStatus = DeploymentStatus.DEPLOYEDFAILED;
-                deployOperation.listEvents.add(new BEvent(EventErrorAtDeployment, e, "Process " + process.getName() + "/" + process.getVersion()));
+                deployOperation.listEvents.add(new BEvent(EventErrorAtDeployment, getLoginformation(process)));
             } catch (ProcessDefinitionNotFoundException e) {
                 deployOperation.deploymentStatus = DeploymentStatus.DEPLOYEDFAILED;
-                deployOperation.listEvents.add(new BEvent(EventErrorAtEnablement, e, "Process " + process.getName() + "/" + process.getVersion()));
+                deployOperation.listEvents.add(new BEvent(EventErrorAtEnablement, e, getLoginformation(process)));
             } catch (ProcessEnablementException e) {
-                deployOperation.deploymentStatus = DeploymentStatus.DEPLOYEDFAILED;
-                deployOperation.listEvents.add(new BEvent(EventErrorAtEnablement, e, "Process " + process.getName() + "/" + process.getVersion()));
+                deployOperation.deploymentStatus = DeploymentStatus.LOADED;
+                deployOperation.listEvents.add(new BEvent(EventErrorAtEnablement, getLoginformation(process)));
             } catch (AlreadyExistsException e) {
                 deployOperation.deploymentStatus = DeploymentStatus.DEPLOYEDFAILED;
-                deployOperation.listEvents.add(new BEvent(EventErrorAtDeployment, e, "Process " + process.getName() + "/" + process.getVersion()));
+                deployOperation.listEvents.add(new BEvent(EventErrorAtDeployment, "Already exist " + getLoginformation(process)));
             }
         }
         return deployOperation;
     }
+
+    private String getLoginformation(Artifact process) {
+        return "ProcessName[" + process.getName() + "] Version[" + process.getVersion() + "]";
+    }
+
 }
